Application Finale

+++ gui/gui_resoudre.h
@@ -8,7 +8,7 @@
 *********************************************************************************************************************************************/
 void choisirFichier(GtkWidget *widget, gpointer data);
 void dialogBoxChoixFichier(GtkWidget *widget, gpointer data);
-void affiche(int grille1[9][9], int grille2[9][9], int tps, int niv,int nbSolution);
+void affiche(int grille1[9][9], int grille2[9][9], float tps, int niv,int nbSolution);
 void retourMenu();
 #endif
